如果下载成功则第二次启动不进行下载;

+++ app/src/main/java/com/gehj/okhttp_netframe/db/DownloadEntity.java
@@ -13,8 +13,7 @@ public class DownloadEntity extends LitePalSupport {
     public long getId() {
         return id;
     }
-
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
